#1516 로봇 조종 테스팅
소켓이 안닫히는 오류 수정

+++ test_ws/StringClassify/RobotControl/RobotControl_move.py
@@ -13,6 +13,12 @@ while True:
     data1 = cli.recv(1024)
     data1 = data1.split(":::")[0]
     print data1
+    if data1 == "cancel":
+        print "sock close"
+        cli.close()
+        sock.close()
+        break
+
     if data2 != data1:
         path = "kill -9 "+str(p.pid)
         p.send_signal(signal.SIGINT)
